Enzo:
EnzoMethodGrackle:

New functions to initialize chemistry fields to fixed (non-zero) fractions of
density, as opposed to the default which would set them all to the same mass
(i.e. not the same fraction). Also sets H and He fields to the sensible
default values.

New function to re-scale all chemistry fields in a given zone if there is
a change in density, provided the new and old density values.