Had to make the static method, EnzoMethodGrackle::define_required_grackle_fields, conditionally compiled when the CONFIG_USE_GRACKLE macro is defined.
Originally we tried defining a version of the function regardless of whether the CONFIG_USE_GRACKLE is defined (but make the internals of the function depend on the presence of the macro). However, when we compiled the program without grackle, we got a seemingly unrelated error complaining about an undefined reference to vtable in the EnzoMethodGrackle constructor. Specifically, the error pointed to the line in member initializer list that called `Method()`.