Cello+Enzo: Allowing methods to self-define their required fields to             prevent user from having to define these fields.
Addition of define_fields and define_group_fields to Methods class, and the
member required_fields_, which is intended to be a vector of field names required
by each method. Intended use is to define required fields using the
required_fields_ member in the initialization of each method, allowing for the
use if if statements for conditional fields depending on run-time parameters
(e.g. as is done in Grackle), then have a call to define_fields to check if fields
are defined, and define them (as permanent) if they do not exist.

define_group_fields ensures that any of the fields (defined or not) are grouped
properly if relevant. The only example of this is in Grackle, which ensures that
the passive tracer species fields are correctly grouped as colour fields for the
hydro solver.

Doing this required re-arranging the order of operations in control_initialize
to initialize methods before the initialization of fields (if the parameter file
initializer was used).