Refactored EnzoIntegrableUpdate to not use EnzoAdvectionFieldLUT
This involved modifying EnzoCenteredFieldRegistry so that it could more easily and directly return details about quantities specified in FIELD_TABLE. To help facilitate this, EnzoCenteredFieldRegistry now constructs and stores a representation of FIELD_TABLE in memory when it is constructed. Additionally, the passive_scalar_group_names instance method was turned into a static method so that the in-memory representation doesn't need to be constructed every time the known names of passively advected scalar groups are needed.