Replaced EnzoAdvectionFieldLUT with EnzoRiemannLUTWrapper
This commit is a substantial overhaul in the implementation of EnzoRiemannImpl and the derived Riemann Solvers. Therefore, significant detail is provided. These details are organized into a description of the context of the change, the motivation for why the change was made, a description of the how solution works, and a brief overview of what actually changed.

Context
=======
This commit overhauls the lookup table (LUT) used internally by EnzoRiemannImpl to map quantity names to array indices. The LUT is used repeatedly within the innermost loop of the Riemann Solver, which is the bottleneck in most hydrodynamical integrators. It is essentially a compromise between the typical solution of defining global macros mapping variable names to indices at compile time (which can be difficult to extend) and a full hash table class (which carry some overhead at runtime). It also allows for the use of a for-loop to uniformly apply an operation for each actively advected quantity that involves multiple values associated with that quantity (e.g. computing flux in the HLL Riemann Solver).

The previous approach made use of EnzoAdvectionFieldLUT. This was a struct with members corresponding to each of the quantities stored in the FIELD_TABLE X MACRO. At runtime, a method of EnzoCenteredFieldTable was used to assign indices to the members of EnzoCenteredFieldRegistry based on a vector of actively advected quantities and then an arrays used to hold values associated with these quantities were dynamically allocated off of the stack. While this approach worked, it to hold these quantities. When a potentially actively advected quantity (e.g. magnetic fields in a pure hydro simulation, internal energy in a simulation without the dual energy formalism) the associated member of the EnzoAdvectionFieldLUT have their values set to -1.

As an example of how this was used, imagine we have an initialized instance of EnzoAdvectionFieldLUT called `lut` and an array called fluxes. To access the entry associated with density or the ith velocity component, one would call `fluxes[lut.density]` or `fluxes[lut.velocity_i]`.

Motivation
==========
While this approach works, there are a couple of drawbacks to the dynamic initialization of the lut. These are highlighted below.
     - The dynamic initialization makes it makes the usage of
       EnzoAdvectionFieldLUT somewhat more difficult to understand
     - the dynamic initialization means that the compiler had less information
       to make optimizations at compile time (compared to the traditional case
       where the mapping of quantity names to array indices are macros).
     - Because the Riemann Solver is a bottleneck, its implementation must make
       a fixed assumptions about the present actively advected quantities.
       Dynamically determining this at runtime would introduce branching. Thus,
       dynamic initialization doesn't actually provide much value.
These considerations make it clear that it would be better to have each Riemann Solver implementation statically specify the LUT at compile time.

Another motivating factor for this change was the way that the `ImplStruct` template arguments was used by EnzoRiemannImpl. The idea was to have `ImplStruct` be a class that provided a function/method to be called in the inner-most loop of EnzoRiemannImpl to perform calculations specific to different Riemann Solvers. To maximize potential future compatibility with the Kokkos framework, it's also important to make sure that function didn't modify any type of internal dynamic state (during the initial implementation, this requirement was also set due to misconceptions about how smp-mode worked). Because the actively advected fields are determined at runtime, the most sensible way to implement this was to require that each struct provide 2 static methods:
    - one method computed the amount of required scratch space based on
      which actively advected quantities are in use (this was necessary to
      dynamically allocate the memory ahead of time before entering the loop)
    - the other method actually performed the solver-specific calculations
However, if the actively advected fields are known at compile time, the requirements on ImplStruct can be drastically simplified. Any scratch space can now be easily allocated off the stack and ImplStruct can now be an idiomatic functor that provides a const `operator()` method. (This is really just an added bonus of overhauling the LUT - it could feasibly be achieved without doing so, but it would introduce new complexity somewhere else).

Solution
========
The solution to this problem is as follows:
  - ImplStruct is a trivially constructible functor that provides a const
    `operator()` instance method that performs solver-specific calculations
    and includes a public declaration of a `LUT` type within the scope of the
    class (via the `using` or `typedef` keyword). The `LUT` type must be a
    specialization of the `EnzoRiemannLUTWrapper<InputLUT>` template class.
  - `InputLUT` is type that provides the mappings of a lookup table with named
    constants. The constants should be named for the subset of the actively
    advected quantities in FIELD_TABLE that are used by a given solver and
    should have values corresponding to unique, contiguous indices in an array.
    The constants corresponding to conserved quantities should all have
    values smaller than the specific quantities. The type must also provide a
    constant, `NEQ`, that indicates the total number of SCALAR quantities or
    VECTOR quantity components included in the table, and a constant
    `specific_start` which specifies the minimum index associated with a
    constant. The type can be implemented with either an unscoped enum OR a
    class which includes some combination of publicly accessible unscoped enums
    and static constant member variables
  - The `EnzoRiemannLUTWrapper<InputLUT>` template class wraps `InputLUT` and
    makes all of its required constants available. It also defines all
    quantities from FIELD_TABLE that did not appear in `InputLUT` and gives
    assigns them values of `-1` (to make it possible to test whether they are
    included in the table). Additionally it provides a handful of helpful
    static methods.

This approach is basically a scoped version of the traditional approach where the mapping of actively advectected quantities to indices are globally defines with macros/enums.

Crude benchmarks indicate a ~25-30% speed increase

The actual changes include
  - removal of all references and methods explicitly related to
    EnzoAdvectionFieldLUT
  - introduction of the EnzoRiemannLUTWrapper template class
  - refactoring of EnzoRiemannImpl and all related components to make use of
    EnzoRiemannLUTWrapper.

The refactored solvers passes the
  - HD_linear_wave_test
  - shock_tube_test
  - dual_energy_shock_tube_test
  - dual_energy_cloud_test
  - passive_advect_sound_test
Also qualitatively confirmed that the Orszag-Tang test yields comparable results to those depicted in Figures 22 and 23 of the athena method paper (Stone08+)

They also passed all of the linear_wave tests except for two cases:
   - the slow wave with N=32
   - comparison of the L1 error norms of the left and right propagating fast
     magnetosonic wave are compares.
The former failure arises because the L1 error norm got smaller after the refactoring and the latter test has never actually passed.

The expected L1 error norms for all of the linear wave tests were updated (even though it was only strictly necessary for the one slow-wave case).