Fixed a bug that had prevented the use of the grackle cooling timestep.
The main issue was that EnzoMethodGrackle's timestep() instance method, declared in src/Enzo/enzo_EnzoMethodGrackle.hpp and implemented in src/Enzo/enzo_EnzoMethodGrackle.cpp was not const-qualified. In contrast, the virtual method declared in the Method super-class (the one Cello uses to compute the timestep) is const-qualified. As a result, the old implementation had been providing an unused overload of the timestep method. When Cello was calling the timestep() for an instance of EnzoMethodGrackle, it the default implementation of Method::timestep was used, which simply returned std::numeric_limits<double>::max().

This bugfix can be tested by adding setting the Method::grackle::use_cooling_timestep parameter to true in input/method_grackle.in. Prior to this commit, changing this value will have no effect on the number of cycles that the simulation will run for. When the changes from this commit are included and the parameter is set to true, then the simulation will run for many,many more cycles.