Refactors calls to the external Grackle library to resolve SMP-mode problems.
The current procedure for setting up the `chemistry_data` struct on each PE (Processor Element) can be summarized by the following 2 steps:

  1. `EnzoConfig` dynamically allocates a local `chemistry_data` struct on each PE (Processor Element). It then calls `set_default_chemistry_parameters`, which both assigns this struct the default parameters and sets the global variable `grackle_data` equal to the address of this struct. Finally, the parameter file values are assigned to the appropriate fields of the struct that is pointed to by `grackle_data`. Note that the `EnzoConfig` destructor will deallocate the locally allocated struct.

  2. The constructor for `EnzoMethodGrackle` calls the instance method `EnzoMethodGrackle::initialize_grackle_chemistry_data`. This method ultimately executes  `initialize_chemistry_data`, which makes some modifications to the struct pointed to by the `grackle_data` global variable and sets up the `grackle_rates` global variable.

When running Enzo-E in non-SMP mode, each PE is on a separate 'logical' node. Thus has its own own copy of the global `grackle_data` and `grackle_rates` variables. However, in SMP-mode there are multiple PEs per 'logical' node and the global `grackle_data` and `grackle_rates` variables are shared between all PEs on a logical node.

This can leads to a few issues in SMP-mode:

  1. A small amount of memory can leak in `EnzoConfig` when multiple PEs try to store the result of `strdup` in `grackle_data->grackle_data_file`.

  2. There are memory leaks caused by the simultaneous calls to `initialize_chemistry_data` (pointers stored in `grackle_rates` that were allocated by one PE may be overwritten by the pointers allocated by another PE)

  3. It seems to be possible for the program to completely crash if the HDF5 files are simultaneously read by different PEs during `initialize_chemistry_data`. This is recorded in Issue #37 (I haven't been able to replicate it since I first reported it - so it seems like an infrequent occurance).

To resolve these issues, I have refactored the code to make use of the explicitly thread-safe local functions from grackle. In doing so, I generally tried to minimize changes to the existing functionallity. The most significant changes are as follows:

  - `EnzoMethodGrackle` now has a new private member variable, `grackle_rates` which stores the local instance of `chemistry_data_storage`. This is initialized in `EnzoMethodGrackle::initialize_grackle_chemistry_data`.

  - `EnzoComputePressure`, `EnzoComputeTemperature`, and `EnzoComputeCoolingTime` were refactored to call the new public instance methods of `EnzoMethodGrackle`: `calculate_pressure`, `calculate_temperature` and `calculate_cooling_time`. Doing this was cleaner than having each compute object retrieve the PE's local instances of `chemistry_data` and `chemistry_data_storage` from `EnzoConfig` and `EnzoMethodGrackle`. This change also makes `EnzoMethodGrackle` better encapuslate the majority of grackle-related operations. To help facillitate these changes, I introduced the `enzo::grackle_method` function to return a pointer to the PE's existing instance of `EnzoMethodGrackle`.

Additionally, I moved the call to `EnzoMethodGrackle::initialize_grackle_chemistry_data` from the `compute` method to the `pup` method (during unpacking). Functionally, this only changes when the grackle structures are initialized after the object has been unpacked. This change removes all chances for the Compute objects to be called immediately after a (hypothetical) restart but before just means that the grackle_data can be initialized.

I have confirmed that this commit does not change the results of executing `input/method_grackle.in` (with just 1 PE or multiple PEs).

Note: I had also considered (and partially implemented) an approach that used locks (`CmiNodeLock`)  and/or barrier synchronization (`CmiNodeBarrier`) to manage access to the global data structures. I ultimately choose to use the thread-safe local functions because it leads to simpler, less fragile code, has the same behavior in SMP and non-SMP mode, and is generally more consistent with the way other Methods are implemented. If there is a strong preference for using the global functions, I can revisit the approach.

In the future, (if these changes are approved) it may be worth creating a physics object to encapsulate all grackle operations (or maybe more generally to encapsulate EquationOfState operations)