Simplified the parts of EnzoConfig::pup that is related to Grackle. Also modified the pup function for chemistry_data to never call strlen on chemistry_data's grackle_data_file field (to avoid issues if the field had not been explicitly set to NULL).
Additionally, the call to strdup to allocate a copy of the string specified by the "Method:grackle:data_file" are swapped out for explicit calls to both new[] and strcpy. Internally, strdup had called malloc. The way EnzoConfig deallocates this memory was also changed from using free to using delete[]. All of this was done to be consistent with the use of new[] to allocate this same pointer in chemistry_data's pup function. (Alternatively, the call to new[] in the pup function could have just been swapped out for malloc and everything else could have stayed the same - but comments would have been required to justify why free and malloc were being used instead of new and delete).