Converted the tests so that they could be run with the master branch of yt
Previously they were run with a custom branch of yt with modifications that have not yet been pushed to master that automatically computed pressure from total_energy, density, velocity and bfields. To make these tests run with the master branch of yt, enzo-e's capibility to compute the pressure field when writing out outputs was used.

All of the tests passed. But for consistency, the exact answers expected for the following tests were updated:
    - run_HD_linear_wave_test
    - run_linear_wave_test
    - run_shock_tube_test
The expected answers for the other vlct tests did not change.