Resolved an issue that caused a segmentation during the checkpoint test
The cause of the issue was a call to InitialValues's initialize_values_ helper method from its pup method while the object was being unpacked. Specifically, initialize_values_ called cello::field_descr()->field_name(index_field) when cello::field_descr had not yet been unpacked. As a consequence cello::field_descr() attempted to access an element of empty vector attribute.

The reason initialize_values_ had been called during unpacking is because the Value class can't currently be pupped. To solve this problem, InitialValue has been refactored so that its values_ attribute is lazily initialized during the first call to enforce_block. Even if unpacking were not an issue this makes a lot of sense because it is unlikely that values will be need to be initialized both before and after checkpointing.