Added the capability for initializing uniform magnetic fields in EnzoInitialCloud. Previously this would have relied on invoking the "value" initializer beforehand (which was not threadsafe). Doing this allows for initializing an MHD simulation of the cloud-crushing problem in smp-mode.
I took some care to test this new way to test the initialization of bfields and confirmed the following 2 facts:
  1. Initializing uniform bfields via EnzoInitialCloud gives the same results between smp and non-smp mode.
  2. Pre-initializing uniform bfields with the value problem initializer leads to the same field values both before AND after this commit.

I also found that for non-trivial initial bfields both means of initializing magnetic fields yield slightly different field values. Expanding on this, the new method implemented in this commit gives the exact expected value while the old method (initializing though the Value initializer) was slightly off (I speculate that the cause for this discrepancy might be a cast of the bfield values in the value initializer before they are used to actually initialize fields).