Refactor EnzoMethodMHDVlct::pup to directly pup the primitive_group_ and bfieldi_group_ Grouping data
This is a necessary change to get the checkpoint restart to work with the MHD integrator. The previous solution to reconstructing the Groupings does not work because it relies on FieldDescr which is not unpacked until after the Method objects.

On a more general note, there isn't really a reason to make primitive_group_ and bfieldi_group_ pointers. They are only pointers due to my inexperience with c++ when I started implementing the method. They would be better off as normal Grouping objects, but making this change has far reaching consequences. This is not addressed in this commit since there are plans to convert all usage of Grouping within VL+CT integrator to some kind of mapping