Refactored EnzoCenteredFieldRegistry to be stateless - This involves the one time construction of an in-memory FIELD_TABLE at startup. - All methods of EnzoCenteredFieldRegistry are now static methods and the constructor is deleted (It's now effectively a glorified namespace). - Several unused/unnecessary methods were deleted. Overall, the implementation is much simpler and easier to follow.
Extra Notes:
------------
Currently we use std::map so we can effectively search for field names. Other types of maps might be better (boost's flat_map would plays better with the cache). However, this is not a performance critical component of the code. Most of this functionallity is just used during setup.

With that in mind, there is definitely a case to be made to switch from using a map with std::string-struct pairs, to just using a std::array containing structs(that additionally contain c-style strings). If we did this, we would basically perform linear searches whenever we query the table. But the tradeoff of this is that there is NO overhead when the program starts up.

The approach we are currently using was chosen because it involved the least number of changes.