Refactored EnzoMethodMHDVlct so that all passively advected scalars don't need to be known at construction.
In general, the fields holding passively advected scalars may not all be initialized until after all the Method objects are initialized (one could also imagine that the initializers may also introduce new fields). This was not completely obvious to me when I initially designed EnzoMethodMHDVlct.

This commit introduced the EnzoNestedPassiveScalarFieldList class. Instances of this class initializes and caches the list of passively advected scalars when they are first needed. Although we don't technically NEED a separate class to do this (this is a conceptually simple idea) I think it does helps reduce the complexity of EnzoMethodMHDVlct. Furthermore, this same functionallity will probably need to be applied to EnzoMethodPpm.

As part of this commit, we also modified EnzoRiemann's factory method and EnzoIntegrableUpdate's constructor so that they don't need to be provided with a list of all passively advected scalars. Earlier changes that cause the list of passively advected scalars to be supplied to instances of these objects during their execution made this relatively straight-forward.