Pruned EnzoFieldArrayFactory.
The shift from using groupings of field names to maps of arrays in EnzoMethodMHDVlct made many of EnzoFieldArrayFactory's methods unnecessary. We removed all of the unused methods and consolidated the remaining methods together.

Made 2 minor improvements to `EnzoFieldArray::from_name`:
- Made the argument a const reference so that the string isn't copied during every call
- Added a check to confirm that the field has the expected precision

In a future PR, we should shift this functionality into Cello's Field class.