Added the CelloArray unit tests to the testing framework
This unit test was originally created and added to the testing framework in the VL+CT pull request. However, while merging that PR and the testing overhaul PR, this test fell through the cracks. This PR rectifies the situation.

I also made a small tweak the unit test itself to address complaints from the compiler that the use of `typename` in nested templates are not compilant with C++11.