Introduced tools/ckpt_restart_test.py and used it to introduce new Checkpoint-Restart tests
This is a command line tool designed to assist with running checkpoint-restart tests. I choose to implement this because many of the PUP routines have minimal/no test coverage. Currently, only ckpt-restart tests only exist for PPM and PPML. While those tests are very useful, they require a lot of boiler plate to setup and would be cumbersome to repeatedly set up.

The tool takes a Cello/Enzo-E input file, a value called RESTART_CYCLE, and a value called STOP_CYCLE. Then it performs the following procedure:

  1. The specified Cello/Enzo-E input file is copied and modified such that the simulation writes a checkpoint output to disk at RESTART_CYCLE, runs until STOP_CYCLE and writes a data output including all fields at STOP_CYCLE.

  2. The simulation is run using the modified input file and all outputs are moved to a separate directory (a copy of the checkpoint is left in the original directory).

  3. The simulation is then restarted and the new outputs are moved to a separate directory.

  4. Finally, the program ensures that the output files written at STOP_CYCLE contain identical data before and after the restart. This is done in a way that accounts for the fact that block data can be stored to hdf5 files in an arbitrary order.

The program writes the result to an output file that can be parsed by the testing framework (this is implemented by tools/test_report.py). This program is slightly more complicated than it needs to be, but it get's the job done. Once PR # 46 is pulled in, we can make the VL+CT tests reuse some of the infrastructure introduced in this commit.

Note that this tool does require that the h5diff utility is installed. I tweaked .circleci/config.yml to make sure that this tool is installed.

I waffled a little between having SCons directly invoke the tool (as a python function) or call it as a command line tool. I choose the latter so that the test is easy to re-run without invoking the entire test suite.

I introduced the following 3 input files and tweaked tools/Restart/SConscript to automatically use them in tests.
- input/Checkpoint/checkpoint_vlct.in
- input/Checkpoint/checkpoint_boundary.in
- input/Checkpoint/checkpoint_grackle.in

Some extra changes were necessary to facillitate the automated running of `checkpoint_grackle.in` (because the path to the Grackle data files depends on the system configuration). In short, the GRACKLE_INPUT_DATA_DIR environment variable is used to specify the path to the directory holding the Grackle input data files. The test/Restart/SConscript file instructs SCons to use that information to writes a short input file that includes the contents of `input/Checkpoint/checkpoint_grackle.in` and overwrites the path to the data file.