Introduced EnzoBfieldMethod and refactored EnzoConstrainedTransport to be a subclass
The changes in the prior commit where I refactored EnzoConstrainedTransport so that it served as a state machine and could be reused for multiple blocks instead of reinitializing it for every block got complicated. I was finding it a little difficult to separate the state machine logic from the actual Constrained Transport Physics.

As a result, I decided to create the EnzoBfieldMethod abstract base class that implements basically tracks the state machine logic and refactored EnzoConstrainedTransport to be a subclass called EnzoBfieldMethodCT. Part of the motivation for doing this was to conform with the style of the rest of the hydro infrastructure (e.g. the Riemann Solvers, Reconstructors, and EquationsOfState). This was also discussed in the Review of VLCT Pull Request (# 62), as a way for potentially allowing Constrained Transport with other methods (like Dedner Divergence Cleaning).

To be completely honest, I'm not completely sold on that the Bfield integration method should be implemented in this way. This also involved a lot more boiler plate code than I was expecting. The choice on how to handle Bfields (if at all) seems much more fundamental that the choice of Reconstruction Algorithm or the even the Riemann Solver and Equation of State. As a long-term solution I think a case could be made for specifying the Bfield Integration Strategy as a template argument. However, that would take a fair amount of additional work and for now we are probably better off conforming with the rest of the hydro infrastructure.

This commit also involves some minor changes to EnzoMethodMHDVlct to actually reuse an instance of the refactored EnzoBfieldMethodCT between cycles and for different blocks.

Finally, it also introduces a new section to the website developer documentation explaining the design and usage of EnzoBfieldMethod.