Tweaked CelloArray's behavior regarding shallow copies of non-initialized CelloArrays.
Previously we always raised an error if an unitialized array was passed to CelloArray's copy constructor & copy assignment operator.
This is not strictly necessary, experience just tells us that this is usually symptomatic of issues (we may need to relax this in the future).
However this caused problems for code resembling the following snippet:

```c++
void func(const bool cond, CelloArray<double,3>* arr_ptr, args...){
  // do some stuff...
  const CelloArray<double,3> arr = (cond) ? *arr_ptr : CelloArray<double,3>();
  // do some other stuff (only touch arr if cond is true) ...
}
```

gcc seemed to handle this idiom pretty well, but icc choked on this.
To rectify this situation, we only raise an error when a shallow copy of an unitialized array is being made AND the destination was never initialized before.

Along the way, I cleaned up a few minor things.