Annotated ``cello::error`` with the ``[[noreturn]]`` attribute
For the following snippet of code:

```
int foo(char arg){
  if (arg == 'a'){
    return 0;
  } else if (arg == 'b'){
    return 1;
  }
  ERROR("foo", "arg must be 'a' or 'b'");
}
```

The compiler would have previously issued a warning (depending on the compiler flags) saying complaining that we are missing a return statement at end of a non-void function. The same thing would happen if we replaced the if-statement with an exhaustive switch statement.

However, now that we have introduced this attribute, the compiler now understands that we will never reach the end of the `foo` function when we raise an error and it will no longer issue a warning.

More details about the `noreturn` attribute can be found here: https://en.cppreference.com/w/cpp/language/attributes/noreturn