Introduced the CArrCollec template class (with unit tests).
This is meant to be used for implementing EnzoEFltArrayMap in order to abstract over how exactly the arrays in the map are stored. This class currently facillitates arrays to be stores as a vector of arrays (as was done previously) and 4d array. In the future, we'll probably simplify this to support just one storage method. But for now, we're supporting both approaches to let us test out optimizations in the Riemann Solver using a 4D array without transitioning other parts of the code. If we decide to support both approaches, there are some additional optimizations to be made.

CArrCollec was primarily introduced in the Cello layer to make it easier to add some unit tests. If we had defined it in the Enzo-layer, it probably wouldn't be a class-template (instead, it would just work with the enzo_floats).