Refactored the EnzoRiemann library so it's a simple library (not a charm++ module).
This changeset moves away from the PUP framework's automatic scheme for serializing subclasses by taking the following steps:
1. I have modified the static factory method, `EnzoRiemann::construct_riemann` to now take an instance of the `EnzoRiemann::FactoryArgs` struct as a single argument (all of the parameters that the method used to accept as arguments are now members within the struct).
2. A copy of the `EnzoRiemann::FactoryArgs` struct (that was used to construct a given instance) is now stored as a member of the `EnzoRiemann` class (and it's subclasses).
3. I've introduced the global function `void operator|(PUP::er &p, EnzoRiemann*& riemann_ptr)` to implement the charm++ pack / unpack functionality. This function essentially PUPs the `EnzoRiemann::FactoryArgs` struct and while unpacking, it passes that struct to the factory method.

Essentially, this slightly increases the complexity of the EnzoRiemann class while significantly simplifying the build system. It's probably a worthwhile tradeoff since the PUP routine for this class is currently only used for checkpoint/restart (NOT migration), and we are planning to move away from Charm++'s checkpoint/restart system.

An added bonus is that the enzo-layer seems to be marginally less sensitive to changes within the EnzoRiemann library. Specifically, changing the names of the template parameter for EnzoRiemannImpl<ImplFunc> triggers less rebuilding in the Enzo-layer (this is a pretty minimal benefit, since the names of the classes don't change much)