Modified dual-energy-formalism handling in EnzoRiemannImplOld
Now we always compute internal_energy_flux and interface_velocity, even if it's not necessary. This is important for minimizing branching in at least some alternative Riemann Solver implementations (if not all implementations).

I've factored out the logic for pre-allocating setting up the internal_energy_flux and velocity_i_bar_array to facillitate simpler comparisons with other implementations (in the future, we probably want to do this in one place, again). The way we allocate scratch space (for cases where we aren't using the dual energy formalism is fairly crude and could definitely be improved).