Fixed bug in cello::define_field_in_group
The function did not return any values (when it should have been returning the id of the newly created field). This caused a segmentation fault whenever this function was executed while using the default settings for "linux_gnu" (which involve heavy optimizations), with at least some version of gcc (I encountered the issue with version 9.3.0).

The only automated test that executed this function is currently the Checkpoint-Restart test for EnzoMethodGrackle. To be clear, ANY simulation involving Grackle would be affected by this bug.