Modified default compilation mode to not use -ffast-math
It turns out that this was the reason certain summary statistics were varying by more than 0.1% when I was using 2 different machines with different compiler versions (gcc versions 9.4.0 vs 7.5.0). This was not completely obvious because Grackle itself was never compiled with these flags (although it took me longer to determine than it should have). As far as I can tell most of the differences arose in EnzoInitialGrackleTest.

We were already making this change in the CMake PR (#139), anyways, so this shouldn't be controversial. As a result of this change, the following modifications were required for existing tests:
- the expected values for the VL+CT dual_energy_cloud test needed to be updated. This change was previously discussed in the context of PR #139.
- the final time for the PPML tests had to be updated for binaries compiled with single precision
- the cosmology test had to be disabled for binaries compiled with single precision. As is described in Issue #154, it appears the `-ffast-math` had been suppressing an existing bug for this case (the test now crashes without the flag). Note that the test ran perfectly fine when compiled in double-precision.

The double-precision results for the Grackle test also had to be updated (but the single-precision results were basically fine). Note that the tolerances for the double-precision results were calibrated so that the tests would pass if you used gcc versions 7.5.0 or 9.4.0 (they may need to be loosened more in the future).

This commit also added the ability for the field_summary to accept different tolerances for different summary statistics (this is useful since we might want 0 tolerance for the location of a min/max while still applying a tolerance for the min/max itself). It also a runtime warnings raised by the script.